Move plotting + Shapefile/MAT to package extensions
Create four weakdep-gated extensions:
- AIBECSRecipesBaseExt = ["RecipesBase", "Interpolations"]: spatial /
  diagnostic recipes (HorizontalPlane, MeridionalPlane, ZonalPlane,
  VerticalPlane, ZPlot, MiniMap, RatioAtStation, PlotStencil) and the
  exported plot* convenience helpers.
- AIBECSRecipesParametersExt = ["RecipesBase", "Distributions"]: the
  PlotParameter / PlotParameters PDF recipes.
- AIBECSShapefileExt = ["Shapefile", "DataFrames"]: real body of
  GroundWaters.load.
- AIBECSOCIM2_48LExt = ["MAT", "NCDatasets"]: real body of OCIM2_48L.load
  (combined trigger because the loader needs both packages).

In src/, GroundWaters.load and OCIM2_48L.load become friendly stubs
that error with the list of required trigger packages. Each extension
file overrides the stub with the real method when its triggers are
loaded.

src/plot_recipes.jl now contains only `function ... end` stubs and
`export` declarations so the API surface stays the same; methods are
attached when the plotting trigger packages are loaded.

RecipesBase, Interpolations, MAT, Shapefile move from [deps] to
[weakdeps]. Drop the unused top-level `using RecipesBase`,
`using Interpolations`, `using Shapefile` from src/AIBECS.jl.

Pattern matches the SciML standard (e.g. LinearSolve.jl/ext/*).
